+++ internal/net/call/handlers.go
@@ -70,9 +70,9 @@ func (hm *HandlerMap) Set(component, method string, handler Handler) {
 	hm.names[fp] = component + "." + method
 }
 
-// addHandlers adds handlers for all methods of the component with the
+// AddHandlers adds handlers for all methods of the component with the
 // specified name. The handlers invoke methods on the specified impl.
-func (hm *HandlerMap) addHandlers(name string, impl any) error {
+func (hm *HandlerMap) AddHandlers(name string, impl any) error {
 	reg, ok := codegen.Find(name)
 	if !ok {
 		return fmt.Errorf("component %s not found", name)
